Add protection against a pending message being consumed by multiple threads somehow.
I believe there is a situation where multiple threads end up getting the same message as we have seen that in our logs. This is NOT supposed to happen with FOR UPDATE, as it is supposed to get a row level lock on the message in question so it should end up throwing a MySqlErrorCode.LockDeadlock transaction, which we ignore and continue. But somehow it looks like multiple threads grab the same message. Not sure if this is a MySQL bug, or a Percona XtraDB clustering issue or what, but we do seem to be seeing this. So to be 100% sure we are not grabbing a message already being processed in another thread, let's make sure leased_by is always null and bail if it is not.

Still had a problem with my original change in production, but it did seem to get rid of most of them. Super strange, but with this change to use a lock table and to join against it, I am now seeing deadlocks more easily from the multiple handler threads like we should be seeing, and you can successfully remove the FOR UPDATE in the SQL and the many threads will make be grabbing the message at the same time but they deadlock on complete and only one of them succeeds. Its better performance I think with FOR UPDATE though as then the threads block *most* of the time correctly on the select function.

Although you could consider not having FOR UPDATE in there and let all the threads deadlock later, as then they will likely end up not aligned exactly anymore like they were.